SolarSimulatorMonitor/requirements.txt
matplotlib~=3.8.0rc1
pyserial~=3.5
bronkhorst-propar~=1.1.0
numpy~=1.25.2
Pillow~=10.0.0
pause~=0.3
pymodbus~=3.5.2